requirements.txt
# Core Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1

# API & Serialization
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
flask-cors==4.0.0
flask-restx==1.1.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2023.3.post1

# Reporting
# exportlab==0.4.3  # Désactivé car non disponible sur PyPI

# Utilities
python-slugify==8.0.1
email-validator==2.1.0.post1
requests==2.31.0

# Development
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-cov==4.1.0

# ZKTeco
pyzk==0.9  # Version mise à jour pour compatibilité
